docker/requirements.txt
Flask==3.0.3
boto3==1.34.131
botocore==1.34.131
GitPython==3.1.43
